pyproject.toml
[project]
name = "stremio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.11.18",
    "aiosqlite>=0.21.0",
    "beautifulsoup4>=4.13.4",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
